requirements.txt
fastapi>=0.115.0
jinja2>=3.1.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
